Create text input mixin
